fix(services-endorsements-api): Fix for undefined enum tags (#4196)
* Added fix for undefined tags

* Removed console.logs

* Fixed tests
