+++ apps/services/endorsements/api/src/app/modules/endorsement/endorsement.service.ts
@@ -10,10 +10,7 @@ import { isPerson } from 'kennitala'
 import { Endorsement } from './endorsement.model'
 import type { Logger } from '@island.is/logging'
 import { LOGGER_PROVIDER } from '@island.is/logging'
-import {
-  EndorsementList,
-  EndorsementTag,
-} from '../endorsementList/endorsementList.model'
+import { EndorsementList } from '../endorsementList/endorsementList.model'
 import { EndorsementMetadataService } from '../endorsementMetadata/endorsementMetadata.service'
 import {
   EndorsementValidatorService,
@@ -22,6 +19,7 @@ import {
 import { EndorsementMetadata } from '../endorsementMetadata/endorsementMetadata.model'
 import { Op, Sequelize, UniqueConstraintError } from 'sequelize'
 import { ValidationRuleDto } from '../endorsementList/dto/validationRule.dto'
+import { EndorsementTag } from '../endorsementList/constants'
 
 interface FindEndorsementInput {
   listId: string
