+++ apps/services/endorsements/api/src/app/modules/endorsementList/e2e/findByTagsEndorsementList/findByTagsEndorsementList.spec.ts
@@ -2,7 +2,7 @@ import { INestApplication } from '@nestjs/common'
 import * as request from 'supertest'
 import { setup } from '../../../../../../test/setup'
 import { errorExpectedStructure } from '../../../../../../test/testHelpers'
-import { EndorsementTag } from '../../endorsementList.model'
+import { EndorsementTag } from '../../constants'
 
 let app: INestApplication
