+++ apps/services/endorsements/api/src/app/modules/endorsementValidator/validators/uniqueWithinTags/uniqueWithinTagsValidator.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common'
 import { IsArray, IsEnum } from 'class-validator'
 import { ValidatorService } from '../../endorsementValidator.service'
-import { EndorsementTag } from '../../../endorsementList/endorsementList.model'
 import { EndorsementMetaField } from '../../../endorsementMetadata/types'
+import { EndorsementTag } from '../../../endorsementList/constants'
 
 export class UniqueWithinTagsInputType {
   @IsArray()
